voicebot/voice-bot/requirements.txt
streamlit
openai
gtts
transformers
torch
audio_recorder_streamlit
python-dotenv
pydub